# openapi/v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "posterProjects Mobile API",
    "version": "2.0.0",
    "description": "Stateless bearer-token surface at /api/v1, consumed by the posterMobile Flutter client. Tokens are Laravel Sanctum personal access tokens named `mobile` with the `mobile` ability, separate from the web session app and the `/mcp` bearer surface (which uses the `mcp` ability). This document describes the surface as implemented, including the exact JSON error contract. Version 2.0.0 is the Marcos OS cutover: the projects, issues, board-columns, sprints and labels operations were removed (their routes are unregistered and answer the generic 404) and replaced by `Objectives` and `Items`; the path prefix stays `/api/v1`."
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative to the application origin. Paths below already include the full /api/v1 prefix."
    }
  ],
...
  "paths": {
...
    "/api/v1/objectives/{objective}/items/{item}": {
      "get": {
        "operationId": "getItem",
        "tags": [
          "Items"
        ],
        "summary": "Show one item by key",
        "description": "The pinned item detail, with its prerequisites and the items it unlocks (each with key, title and derived state). Items of a closed objective resolve read-only. Serialization does not grow in queries with the number of neighbours. Requires the `mobile` ability.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "objective",
            "in": "path",
            "required": true,
            "description": "The objective's `key` (e.g. SALUD) — never the numeric `id`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item",
            "in": "path",
            "required": true,
            "description": "The item's public key (e.g. SALUD-7): the objective key, a dash and the item number. Resolved in code, never by route-model binding.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The item.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                },
                "example": {
                  "data": {
                    "id": 7,
                    "key": "SALUD-2",
                    "kind": "task",
                    "title": "Captura 10 min",
                    "description": null,
                    "two_minute_version": "abrir el inbox y escribir una línea",
                    "state": "available",
                    "target_date": null,
                    "plan": {
                      "id": 3,
                      "title": "Semana 1"
                    },
                    "prerequisites": [
                      {
                        "key": "SALUD-1",
                        "title": "Mesa lista",
                        "state": "done"
                      }
                    ],
                    "unlocks": [
                      {
                        "key": "SALUD-3",
                        "title": "Clasificar y elegir prioridad",
                        "state": "locked"
                      }
                    ],
                    "completed_at": null,
                    "evidence": null,
                    "updated_at": "2026-09-28T03:00:00+00:00"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No token, or a token that does not authenticate.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge.",
                "schema": {
                  "type": "string",
                  "example": "Bearer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedResponse"
                },
                "example": {
                  "message": "No autenticado."
                }
              }
            }
          },
          "403": {
            "description": "A valid token that lacks the `mobile` ability (e.g. an `mcp`-only token).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                },
                "example": {
                  "message": "Este token no tiene permiso para usar esta API."
                }
              }
            }
          },
          "404": {
            "description": "Any of: a key with no dash, an empty or non-numeric suffix, a prefix that is not `{objective}`'s key, a well-formed key whose number does not exist in this objective, a retired item, or an unresolvable `{objective}`. All identical.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                },
                "example": {
                  "message": "Recurso no encontrado."
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/objectives/{objective}/items/{item}/check": {
      "post": {
        "operationId": "checkItem",
        "tags": [
          "Items"
        ],
        "summary": "Mark an item done",
        "description": "Marks an available or active item `done` through the same domain action as the web and answers the item detail plus `unlocked` (the items that became available). A locked item is `422`; a milestone needs a non-empty `evidence` (`422` otherwise); an already-done item is `200` unchanged (idempotent). Items of a closed objective are read-only (`422`). Requires the `mobile` ability.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "objective",
            "in": "path",
            "required": true,
            "description": "The objective's `key` (e.g. SALUD) — never the numeric `id`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item",
            "in": "path",
            "required": true,
            "description": "The item's public key (e.g. SALUD-7): the objective key, a dash and the item number. Resolved in code, never by route-model binding.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckItemRequest"
              },
              "example": {
                "evidence": "Boceto en la libreta, página 12"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The item, now done, and what it unlocked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemCheckResponse"
                },
                "example": {
                  "data": {
                    "id": 7,
                    "key": "SALUD-2",
                    "kind": "task",
                    "title": "Captura 10 min",
                    "description": null,
                    "two_minute_version": "abrir el inbox y escribir una línea",
                    "state": "done",
                    "target_date": null,
                    "plan": {
                      "id": 3,
                      "title": "Semana 1"
                    },
                    "prerequisites": [
                      {
                        "key": "SALUD-1",
                        "title": "Mesa lista",
                        "state": "done"
                      }
                    ],
                    "unlocks": [
                      {
                        "key": "SALUD-3",
                        "title": "Clasificar y elegir prioridad",
                        "state": "locked"
                      }
                    ],
                    "completed_at": "2026-09-28T03:10:00+00:00",
                    "evidence": null,
                    "updated_at": "2026-09-28T03:00:00+00:00",
                    "unlocked": [
                      {
                        "key": "SALUD-3",
                        "title": "Clasificar y elegir prioridad",
                        "state": "available"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "No token, or a token that does not authenticate.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge.",
                "schema": {
                  "type": "string",
                  "example": "Bearer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedResponse"
                },
                "example": {
                  "message": "No autenticado."
                }
              }
            }
          },
          "403": {
            "description": "A valid token that lacks the `mobile` ability (e.g. an `mcp`-only token).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                },
                "example": {
                  "message": "Este token no tiene permiso para usar esta API."
                }
              }
            }
          },
          "404": {
            "description": "Any of: a key with no dash, an empty or non-numeric suffix, a prefix that is not `{objective}`'s key, a well-formed key whose number does not exist in this objective, a retired item, or an unresolvable `{objective}`. All identical.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                },
                "example": {
                  "message": "Recurso no encontrado."
                }
              }
            }
          },
          "422": {
            "description": "A locked item, a milestone without evidence, or an item of a closed objective.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                },
                "example": {
                  "message": "Esta tarea está bloqueada: se abre al terminar Mesa lista.",
                  "errors": {
                    "item": [
                      "Esta tarea está bloqueada: se abre al terminar Mesa lista."
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum personal access token",
        "description": "Plain-text Sanctum token minted by POST /api/v1/login, sent as `Authorization: Bearer <token>`. Carries the `mobile` ability (or the legacy wildcard `['*']`)."
      }
    },
    "schemas": {
...
      "NotFoundResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Recurso no encontrado."
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "required": [
          "message",
          "errors"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "The first validation error message."
          },
          "errors": {
            "type": "object",
            "description": "Field name to array of error messages.",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "UnauthenticatedResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "No autenticado."
          }
        }
      },
      "ForbiddenResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Este token no tiene permiso para usar esta API."
          }
        }
      },
...
      "ItemRef": {
        "type": "object",
        "required": [
          "key",
          "title",
          "state"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "locked",
              "available",
              "active",
              "done"
            ]
          }
        }
      },
      "ItemDetail": {
        "type": "object",
        "description": "Pinned item detail: exactly these 14 fields.",
        "required": [
          "id",
          "key",
          "kind",
          "title",
          "description",
          "two_minute_version",
          "state",
          "target_date",
          "plan",
          "prerequisites",
          "unlocks",
          "completed_at",
          "evidence",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "key": {
            "type": "string",
            "example": "SALUD-7"
          },
          "kind": {
            "type": "string",
            "enum": [
              "task",
              "milestone"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "two_minute_version": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "locked",
              "available",
              "active",
              "done"
            ]
          },
          "target_date": {
            "type": "string",
            "format": "date",
            "description": "Optional; a past date never changes the state.",
            "nullable": true
          },
          "plan": {
            "type": "object",
            "required": [
              "id",
              "title"
            ],
            "properties": {
              "id": {
                "type": "integer"
              },
              "title": {
                "type": "string"
              }
            }
          },
          "prerequisites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemRef"
            }
          },
          "unlocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemRef"
            }
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "evidence": {
            "type": "object",
            "description": "Milestones only (else null).",
            "properties": {
              "text": {
                "type": "string"
              },
              "link": {
                "type": "string",
                "nullable": true
              }
            },
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "ItemResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ItemDetail"
          }
        }
      },
      "ItemCheckResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ItemDetail"
              },
              {
                "type": "object",
                "required": [
                  "unlocked"
                ],
                "properties": {
                  "unlocked": {
                    "type": "array",
                    "description": "Items that became available because of this check (empty when the item was already done).",
                    "items": {
                      "$ref": "#/components/schemas/ItemRef"
                    }
                  }
                }
              }
            ]
          }
        }
      },
      "CheckItemRequest": {
        "type": "object",
        "properties": {
          "evidence": {
            "type": "string",
            "maxLength": 2000,
            "description": "Required and non-empty for a milestone; ignored for a task.",
            "nullable": true
          },
          "link": {
            "type": "string",
            "format": "uri",
            "maxLength": 2048,
            "nullable": true
          }
        }
      }
    }
  }
}